netbox/config/requirements.txt
pyyaml
pynetbox